[Add support for local file and web server polling modes]
Detailed Changes:
- Added support for reading election results from a local JSON file specified via the --local-file command line argument
- Added support for polling election results from a web server at https://electoral-console.dokku.heare.io/ if --local-file is not specified
- Added the aiohttp library as a dependency to enable the web server polling functionality

Impact: This change adds flexibility to the electoral map visualization application by allowing users to specify a local file for the election results or fall back to polling a web server if no local file is provided. This should improve the usability and accessibility of the application.

Additional Notes: The application now supports two modes of operation: local file polling and web server polling. Users can choose which mode to use by providing the --local-file argument or leaving it unspecified.